Clean up the image pre-processing code
Keras in Tensorflow has a function called 'image_dataset_from_directory', which requires the images to be stored in separate folders depending on their label. This commit, in addition to calling the required pre-processing functions (crop ROI and augmentation), also stores the resulting images in this fashion.

After cropping the ROI, the image must be resized before being passed through the network. I cannot find an equivalent function in the old version of this file (found in the 'master' branch). I could only find a 'reshape' command in the 'plot_augment' funciton definition, which I have now removed. This has been addressed with a 'resize' command in line 80 that makes use of linear interpolation.

Unresolved issues:

- Erosion and Dilation with the current parameters only seem to worsen the image. Experimenation is required to arrive at the optimal set of parameters. These lines (lines 70-72) have been commented for now.

- Unsure if we cropped the ROI before or after augmentation. This sequence has been adopted for now.

- The implications of the 'resize' command detailed above has to be investigated.